01-python-ai-integration/requirements.txt
openai>=2.15.0
python-dotenv>=1.2.1
jsonschema>=4.26.0
